create service added

+++ src/main/java/com/homehero/servicios/DTOServicios/HeroServiceDTO.java
@@ -9,7 +9,7 @@ public class HeroServiceDTO {
 
     private String name;
     private int service_id;
-    private double price;
+    private int price;
     private String description;
     private int hero_id;
